requirements.txt
requests>=2.28.0
pandas>=1.5.0
jieba>=0.42.1
python-dotenv>=1.0.0
openpyxl>=3.0.0